refactor: move chat to lib
