+++ src/lib/chat.ts
